app/requirements.txt
alembic==1.12.0
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
Django==4.2.5
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-cms==3.11.4
django-formtools==2.4.1
django-js-asset==2.1.0
django-sekizai==4.1.0
django-treebeard==4.7
djangocms-admin-style==3.2.6
djangocms-text-ckeditor==5.1.3
docopt==0.6.2
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
html5lib==1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kafka==1.3.5
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
neovim==0.3.1
nltk==3.8.1
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
py4j==0.10.9.7
pyflakes==3.0.1
Pygments==2.15.1
pynvim==0.4.3
pyspark==3.5.0
python-dotenv==1.0.0
pyvim==3.0.3
redis==5.0.1
regex==2023.10.3
rq==1.15.1
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.0
